Send SIGINT to worker processes instead of SIGKILL
This allows us to cleanup the VMAF report. When Go kills a process, it
instantly stops paying attention to it. However, FFmpeg seems to create
the report file in one of its children, which doesn't create the report
until a short while after the kill signal is sent and parent FFmpeg
process ignored. This meant cleanup would happen too soon.

Sending an interrupt allows FFmpeg to terminate gracefully, which means
the report is generated while Wait() is still waiting for the process.