Rearrange in to idiomatic Go project layout
While this project is on the small side to require the layout,
separating the files in to separate packages can help exposes weaknesses
in the API design.

For example, both the conversion and vmaf packages have three very
similar funcs (possible, required, perform) that could be put against an
job interface. Likewise, the flags could be included in such an
interface, as they are referred to and used everywhere.